requirements.txt
environs==9.5.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-api-keys==0.5.3
google-cloud-dialogflow==2.23.2
python-telegram-bot==13.15
vk-api==11.9.9
click~=8.1.7